Fix StreamService.User(): parse very long friend ID lists properly
This is a bug fix for the issue that Stream.Messages of User Stream
returns no message if the user has a large number of friends (say
7,000).

Background:
* go-twitter users bufio.Scanner.Scan() to scan Twitter streaming
messages, but it has a limitation that it can scan lines whose length is
up to bufio.MaxScanTokenSize (65,536) bytes by default. If the line
length of a streaming data exceeds the limit, Scanner will just stop
scanning the stream.
* User Stream sends a list of friend IDs right after establishing a
connection.
* If the user has a large number of friends (say 7,000), the friend ID
list from User Stream will easily exceed bufio.MaxScanTokenSize, and
Scanner will stop scanning streaming messages, and a client will not get
any message.

Solution:
* Use bufio.Reader instead. This has no line length limit.