Drop support for go-twitter library and mark deprecated
* The go-twitter API library will not be developed further
or maintained
* The project will be left as-is to allow users to migrate
to new platforms. Then later archived as read-only

I developed this library in 2015 and have maintained it for the
last 7 years. I'm grateful to the 44 contributors and 1 sponsor
the project has had at various points. Twitter was a wonderful
platform, built by a wonderful team. With its recent changes,
its time to re-evaluate what we invest in and the world we want
to build. To the 1400+ projects using this library, I'll leave
the repo read-only so you can migrate.

All the best,
Dalton